Changes done to SecurityConfig
* No one tested that class and just decided to push to main.
So I corrected that code and actually ran the application as well as testing on Postman.

Next time check your stuff before pushing to main...
