+++ src/main/resources/application.properties
@@ -4,6 +4,6 @@ spring.datasource.username=root
 spring.datasource.password=
 spring.jpa.hibernate.ddl-auto=update
 
-#security
+#security - Not needed as we have a security package
 #spring.security.user.name=user
 #spring.security.user.password=password
